Updates the template to include a local host bind
We dont wish to expose ETCD to the public interface, however the etcd
utlity appears to default to wanting to communicate over 127.0.0.1:4001.
It does expose the -C and --peers option to reconfigure itself to
communicate with a peer on the given ip:port combination, but this
breaks user expectation of things "just working".

This fix adjusts the binding which can be verified via netstat -nlp that
hte service is only listening on private address, and the adveritsing to
the API/Peers only displays the private address, while allowing the
etcdctl client to communicate using the localhost binding default.